Factoring out a GabrielServer base class and having the zmq/websockets implementations inherit from it.
